feat!: expose Prettier config, name config functions consistently
